+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Fixed
 
+- Fix validating empty, /* and /** channel names.
+- Fix subscribing to multiple channels.
+- Fix wrong channel name in client message.
+
 ### Add
 
 - Channel name validation.